+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.0.4] - January 3rd, 2024
+
+* Updated to be compatible with `web_socket_channel` version `2.4.1`.
+* Preparing to archive package as it's no longer necessary since `web_socket_channel` version `2.3.0`
+
+
 ## [1.0.3] - November 15th, 2023
 
 * Standardized analysis rules
@@ -86,6 +92,7 @@
 ## [1.0.0] - November 12th, 2022
 
 * Initial release
+